Add files via upload
It was about the crime reporting in India between years of 2001-2010 in five major states of India. The data set was analyzed and brought up to the interactive dashboard which communicate and tell a story about crimes in India , how many were reported , how many were punished , how many males and females were victims , difference of males and female numbers of crimes , which crimes were highly reported and committed.